refactor commit messages

+++ routes/pasteRouter.js
@@ -18,39 +18,43 @@ routes.get(['/:documentId', '/:slug/:documentId', '/:slug/:slugTwo/:documentId',
         fs.readFile(`./pastes/${documentId}.txt`, 'utf-8', (err, code) => {
             if (code) {
                 db.link.findOne({ URL: documentId }, (err, doc) => {
-                    var enableImageEmbed;
-                    if (!doc.instantDelete) {
-                        const month = new Date(doc.deleteDate).getMonth() + 1 // we have to put a + 1 here because months start at 0
-                        const day = new Date(doc.deleteDate).getDate()
-                        const year = new Date(doc.deleteDate).getFullYear()
-                        var deleteDate = `Deletes on ${month}/${day}/${year}`
-                    } else {
-                        if (!req.isCrawler()) {
-                            setTimeout(() => {
-                                fs.unlink(`./pastes/${doc.URL}.txt`, err => {
-                                    if (err) return err;
-                                    db.link.remove({ _id: doc._id })
-                                })
-                            }, 1000);
-                        }
-                        var deleteDate = 'Deletes after being viewed.'
-                    }
-                    if (doc.imageEmbed && fs.existsSync(`./public/assets/img/${documentId}.jpeg`))
-                        var enableImageEmbed = true
-                    else
-                        var enableImageEmbed = false
-                    if (req.isAuthenticated()) {
-                        Users.findOne({ _id: req.user }, (err, user) => {
-                            const editorArray = doc.allowedEditor
-                            if (req.user.toString() == doc.creator || editorArray.indexOf(req.user.toString()) != -1) {
-                                var creator = true;
-                            } else {
-                                var creator = false;
+                    if (doc) {
+                        var enableImageEmbed;
+                        if (!doc.instantDelete) {
+                            const month = new Date(doc.deleteDate).getMonth() + 1 // we have to put a + 1 here because months start at 0
+                            const day = new Date(doc.deleteDate).getDate()
+                            const year = new Date(doc.deleteDate).getFullYear()
+                            var deleteDate = `Deletes on ${month}/${day}/${year}`
+                        } else {
+                            if (!req.isCrawler()) {
+                                setTimeout(() => {
+                                    fs.unlink(`./pastes/${doc.URL}.txt`, err => {
+                                        if (err) return err;
+                                        db.link.remove({ _id: doc._id })
+                                    })
+                                }, 1000);
                             }
-                            res.render('pasted.ejs', { documentName: documentId, imageEmbed: enableImageEmbed, code: code, loggedIn: true, pfp: user.icon, deleteDate: deleteDate, creator: creator, originalCreator: doc.creator, incomingUser: req.user.toString() })
-                        })
+                            var deleteDate = 'Deletes after being viewed.'
+                        }
+                        if (doc.imageEmbed && fs.existsSync(`./public/assets/img/${documentId}.jpeg`))
+                            var enableImageEmbed = true
+                        else
+                            var enableImageEmbed = false
+                        if (req.isAuthenticated()) {
+                            Users.findOne({ _id: req.user }, (err, user) => {
+                                const editorArray = doc.allowedEditor
+                                if (req.user.toString() == doc.creator || editorArray.indexOf(req.user.toString()) != -1) {
+                                    var creator = true;
+                                } else {
+                                    var creator = false;
+                                }
+                                res.render('pasted.ejs', { documentName: documentId, imageEmbed: enableImageEmbed, code: code, loggedIn: true, pfp: user.icon, deleteDate: deleteDate, creator: creator, originalCreator: doc.creator, incomingUser: req.user.toString() })
+                            })
+                        } else {
+                            res.render('pasted.ejs', { documentName: documentId, imageEmbed: enableImageEmbed, code: code, loggedIn: false, deleteDate: deleteDate, creator: false })
+                        }
                     } else {
-                        res.render('pasted.ejs', { documentName: documentId, imageEmbed: enableImageEmbed, code: code, loggedIn: false, deleteDate: deleteDate, creator: false })
+                        res.render('error.ejs', { error: 'We couldn\'t find that document!' });
                     }
                 })
             } else if (err) {
